task: Generic NftCard and new NftCardWithHover component created
