+++ components/base/NftCard/index.tsx
@@ -1,2 +1,3 @@
 export { default } from './NftCard';
+export { default as NftCardWithHover } from './components/NftCardWithHover';
 export type { ModeType } from './NftCard';
